# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Proxmaster API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check with fail-closed gate state
      responses:
        "200":
          description: OK
  /cluster/overview:
    get:
      summary: Cluster overview
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cluster state
  /storage/inventory:
    get:
      summary: Get unified storage inventory (pools, datastores, tiers, replication, targets)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Storage inventory
  /storage/rebuild/plan:
    post:
      summary: Create rebuild-all-pools plan (guarded)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rebuild plan
  /backup/policies:
    get:
      summary: List backup policies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policies
    post:
      summary: Upsert backup policy for a workload
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy upserted
  /backup/targets:
    get:
      summary: List backup targets (PBS/NFS/S3)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Targets
  /jobs:
    get:
      summary: List jobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Jobs
  /incidents:
    get:
      summary: List incidents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Incidents
  /policy/simulate:
    post:
      summary: Simulate policy decision for a tool call
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tool:
                  type: string
                params:
                  type: object
                approve_now:
                  type: boolean
      responses:
        "200":
          description: Simulation result
  /policy/explain:
    get:
      summary: Explain active policy mode and gates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy explanation
  /controlplane/endpoint:
    get:
      summary: Get currently active control-plane endpoint (VIP/DNS handover target)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Endpoint details
  /connectivity/status:
    get:
      summary: WireGuard, GitOps and break-glass connectivity status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connectivity status
  /gitops/status:
    get:
      summary: Get GitOps sync/deploy status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: GitOps status
  /gitops/sync:
    post:
      summary: Trigger GitOps sync now
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sync job response
  /gitops/rollback:
    post:
      summary: Roll back to last known stable commit
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rollback job response
  /access/breakglass:
    get:
      summary: Get break-glass SSH status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Break-glass status
  /access/breakglass/enable:
    post:
      summary: Enable break-glass SSH (time-boxed)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Break-glass enable response
  /access/breakglass/disable:
    post:
      summary: Disable break-glass SSH
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Break-glass disable response
  /vpn/wireguard/status:
    get:
      summary: Get WireGuard status via MCP
      security:
        - bearerAuth: []
      responses:
        "200":
          description: WireGuard status
  /vpn/wireguard/plan:
    post:
      summary: Generate WireGuard setup plan and client config template
      security:
        - bearerAuth: []
      responses:
        "200":
          description: WireGuard plan
  /vpn/wireguard/apply:
    post:
      summary: Apply WireGuard server config on Proxmaster VM (guarded)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: WireGuard apply response
  /mcp/call:
    post:
      summary: Execute MCP tool call
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
          required: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tool:
                  type: string
                params:
                  type: object
                actor:
                  type: string
                approve_now:
                  type: boolean
                idempotency_key:
                  type: string
                hardware_mfa:
                  type: boolean
                second_approver:
                  type: string
      responses:
        "200":
          description: Tool execution result
      description: |
        Includes storage/backup tools:
        `connectivity.status`, `proxmox.connection.test`,
        `gitops.status`, `gitops.sync.now`, `gitops.rollback`,
        `ssh.breakglass.status`, `ssh.breakglass.enable`, `ssh.breakglass.disable`,
        `vpn.wireguard.status`, `vpn.wireguard.plan`, `vpn.wireguard.apply`,
        `storage.inventory.sync`, `storage.pool.plan_apply`, `storage.pool.rebuild_all.plan`,
        `storage.pool.rebuild_all.execute`, `storage.replication.plan_apply`, `storage.health.explain`,
        `backup.policy.upsert`, `backup.policy.explain`, `backup.run.now`,
        `backup.restore.plan`, `backup.restore.execute`, `backup.verify.sample`.
  /mcp/approve:
    post:
      summary: Approve and execute guarded high-risk operation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Approval result
      description: |
        Supports high-risk tools like `proxmaster.self_migrate` for seamless control-plane switch.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT